+++ controlePedidos/src/main/java/com/br/chocolatePontoVirgula/model/entity/ItemPedido.java
@@ -27,9 +27,9 @@ public class ItemPedido {
 	private Double valorTotal;
 
 	@Column(name = "codigo_pedido")
-	private int codigoPedido;
+	private Long codigoPedido;
 
 	@Column(name = "codigo_produto")
-	private int codigoProduto;
+	private Long codigoProduto;
 
 }
